Bugfix/play history rebase (#3)
* resolved rebase issues and converted play tracker to play history

* imports
